+++ CHANGELOG.md
@@ -4,15 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+
+## [Unreleased] 2025-02-11
+
+### Added
+
+- Throw error if input file is empty.
+
+
 ## [1.0.1] 2025-02-11
 
 ### Added
 
-- handle cancelling of workflow
+- Handle cancelling of workflow.
 
 ## [1.0.0] 2024-11-14
 
 ### Added
 
-- initial version
+- Initial version.